+++ docs/changes.rst
@@ -38,6 +38,10 @@ Released: not yet
 
 **Enhancements:**
 
+* Added a 'easy-vault delete-password' command that deletes the password for
+  a vault file in the keyring service. Added a corresponding
+  'Keyring.delete_password()' method. (issues #33 and #35)
+
 **Cleanup:**
 
 **Known issues:**